requirements.txt
streamlit
google-generativeai
feedparser
streamlit-mic-recorder